change: moved `router` under folder apisix.
