+++ lua/apisix.lua
@@ -6,7 +6,7 @@ local plugin = require("apisix.plugin")
 local service_fetch = require("apisix.http.service").get
 local admin_init = require("apisix.admin.init")
 local get_var = require("resty.ngxvar").fetch
-local router = require("apisix.http.router")
+local router = require("apisix.router")
 local ngx = ngx
 local get_method = ngx.req.get_method
 local ngx_exit = ngx.exit
