+++ lua/apisix/admin/upstreams.lua
@@ -1,5 +1,5 @@
 local core = require("apisix.core")
-local get_routes = require("apisix.http.router").http_routes
+local get_routes = require("apisix.router").http_routes
 local get_services = require("apisix.http.service").services
 local tostring = tostring
 local ipairs = ipairs
